Improve banner visibility and display logic
